refactor: Move command specific logic to respective command modules
