+++ pueue/src/client/commands/state/query/column_selection.rs
@@ -1,9 +1,7 @@
 use pest::iterators::Pair;
 
-use crate::{
-    client::query::{QueryResult, Rule},
-    internal_prelude::*,
-};
+use super::{QueryResult, Rule};
+use crate::internal_prelude::*;
 
 pub fn apply(section: Pair<'_, Rule>, query_result: &mut QueryResult) -> Result<()> {
     // This query is expected to be structured like this:
