+++ pueue/src/client/commands/state/query/limit.rs
@@ -1,9 +1,7 @@
 use pest::iterators::Pair;
 
-use crate::{
-    client::query::{QueryResult, Rule},
-    internal_prelude::*,
-};
+use super::{QueryResult, Rule};
+use crate::internal_prelude::*;
 
 /// An enum indicating whether the first or the first or the last tasks in the
 /// `pueue status` command should be shown.
