+++ pueue/tests/client/mod.rs
@@ -1,3 +1,2 @@
 mod helper;
 mod integration;
-mod unit;
