--- pueue/tests/client/unit/mod.rs
@@ -1,1 +1,0 @@
-mod status_query;
